third_party/WebKit/LayoutTests/http/tests/inspector/tracing/timeline-xhr-response-type-blob-event.html

<html>
<head>
<script src="../inspector-test.js"></script>
<script src="../timeline-test.js"></script>
<script src="../network-test.js"></script>
<script>

function performActions(callback)
{
    var xhr = new XMLHttpRequest();
    xhr.responseType = "blob";
    xhr.open("GET", "network/resources/resource.php", true);
    xhr.onload = function() { };  // This is necessary for XHRLoad event.
    // assigning callback to onload doesn't work here due to exception in responseXML handling for blob response type.
    xhr.onreadystatechange = done;
    function done()
    {
        if (xhr.readyState === 4)
            callback();
    }
    xhr.send(null);
}

function test()
{
    InspectorTest.resetInspectorResourcesData(step1);
    function step1()
    {
        InspectorTest.invokeAsyncWithTimeline("performActions", finish);

        function finish()
        {
            InspectorTest.printTimelineRecords("XHRReadyStateChange");
            InspectorTest.printTimelineRecords("XHRLoad");
            InspectorTest.completeTest();
        }
    }
}

</script>
</head>

<body onload="runTest()">
<p>
Tests the Timeline events for XMLHttpReqeust with responseType="blob"
</p>

</body>
</html>
