third_party/WebKit/LayoutTests/http/tests/inspector/tracing/timeline-xhr-response-type-blob-event-expected.txt

Tests the Timeline events for XMLHttpReqeust with responseType="blob"

XHRReadyStateChange Properties:
{
    data : {
        frame : <string>
        readyState : 1
        stackTrace : <object>
        url : .../inspector/tracing/network/resources/resource.php
    }
    endTime : <number>
    frameId : <string>
    stackTrace : <object>
    startTime : <number>
    type : "XHRReadyStateChange"
}
XHRReadyStateChange Properties:
{
    data : {
        frame : <string>
        readyState : 2
        url : .../inspector/tracing/network/resources/resource.php
    }
    endTime : <number>
    frameId : <string>
    startTime : <number>
    type : "XHRReadyStateChange"
}
XHRReadyStateChange Properties:
{
    data : {
        frame : <string>
        readyState : 3
        url : .../inspector/tracing/network/resources/resource.php
    }
    endTime : <number>
    frameId : <string>
    startTime : <number>
    type : "XHRReadyStateChange"
}
XHRReadyStateChange Properties:
{
    data : {
        frame : <string>
        readyState : 4
        url : .../inspector/tracing/network/resources/resource.php
    }
    endTime : <number>
    frameId : <string>
    startTime : <number>
    type : "XHRReadyStateChange"
}
XHRLoad Properties:
{
    data : {
        frame : <string>
        url : .../inspector/tracing/network/resources/resource.php
    }
    endTime : <number>
    frameId : <string>
    startTime : <number>
    type : "XHRLoad"
}
